refactor(api): unify model imports to index.js for consistency across API files
