+++ src/apis/addReaction.ts
@@ -1,6 +1,5 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
-import { ThreadType } from "../models/index.js";
-import { Reactions } from "../models/Reaction.js";
+import { ThreadType, Reactions } from "../models/index.js";
 import { apiFactory } from "../utils.js";
 
 export type AddReactionResponse = {
