+++ src/apis/getAllFriends.ts
@@ -1,5 +1,5 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
-import { Gender } from "../models/Enum.js";
+import { Gender } from "../models/index.js";
 import { apiFactory } from "../utils.js";
 
 export type GetAllFriendsResponse = {
