+++ src/apis/getListBoard.ts
@@ -1,6 +1,5 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
-import { NoteDetail, PinnedMessageDetail, PollDetail } from "../models/Board.js";
-import { BoardType } from "../models/Enum.js";
+import { NoteDetail, PinnedMessageDetail, PollDetail, BoardType } from "../models/index.js";
 import { apiFactory } from "../utils.js";
 
 export type ListBoardOptions = {
